pyproject.toml
[tool.poetry]
name = "chere-bot"
version = "0.1.0"
description = ""
authors = ["Nick <nickphilomath@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
python-telegram-bot = "^21.2"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.1"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
